# Covenant/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Covenant API",
    "version": "v0.1"
  },
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "GetUsers",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CovenantUser"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "CreateUser",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CovenantUserLogin"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CovenantUser"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "EditUser",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CovenantUserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CovenantUser"
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "GetUser",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CovenantUser"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "DeleteUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/users/login": {
      "post": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "Login",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CovenantUserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CovenantUserLoginResult"
            }
          }
        }
      }
    },
...
    "/api/users/{id}/roles": {
      "get": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "GetUserRoles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StringIdentityUserRole"
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/roles/{rid}": {
      "get": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "GetUserRole",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "default": ""
          },
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringIdentityUserRole"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "CreateUserRole",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "default": ""
          },
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringIdentityUserRole"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "DeleteUserRole",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "default": ""
          },
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/roles/{rid}": {
      "get": {
        "tags": [
          "CovenantUserApi"
        ],
        "operationId": "GetRole",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "type": "string",
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/IdentityRole"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CovenantUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "CovenantUserLogin": {
      "required": [
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "CovenantUserLoginResult": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "covenantToken": {
          "type": "string"
        }
      }
    },
    "StringIdentityUserRole": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "roleId": {
          "type": "string"
        }
      }
    },
    "IdentityRole": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "normalizedName": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  }
}